flink-python/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.7
protobuf==3.20.3
google
apache-flink==1.18.0